udpates - renamed edit listng form to processing listing form - removed the create code on the view details link - decoupled create listing from the listing details component - DRY up the containers into one component
