+++ app/javascript/components/Listing/index.js
@@ -10,7 +10,7 @@ import { canShow } from '../../utils/permissions';
 
 // Components
 import ListingDetails from './listing_details';
-import EditListingForm from './edit_listing_form';
+import ProcessListingForm from './process_listing_form';
 
 // TODO : LOOK AT REDIRECT CACHE https://www.apollographql.com/docs/react/caching/cache-interaction/#cache-redirects-with-cacheredirects
 
@@ -38,26 +38,24 @@ const CURRENT_USER = gql`
 `;
 
 // eslint-disable-next-line react/display-name
-export default ({ listingId = '', edit = false, createNewListing = false }) => {
+export default ({ listingId = '', edit = false }) => {
   const client = useApolloClient();
 
   const [editing, setEditing] = useState(edit);
   const [listing, setListing] = useState(listingId);
-  const [creating, setCreating] = useState(createNewListing);
 
   // if the edit prop changes, updates the components edit state to toggle views properly
   useEffect(() => {
     if (editing !== edit) setEditing(edit);
     if (listing !== listingId) setListing(listingId);
-    if (creating !== createNewListing) setCreating(createNewListing);
-  }, [edit, editing, listing, listingId, createNewListing, creating]);
+  }, [edit, editing, listing, listingId]);
 
   const toggleEditMode = e => {
     if (e) {
       // if used with a Link component
       e.preventDefault();
       return client.writeData({
-        data: { selectedListingId: listingId, edit: !editing, createNewListing: !createNewListing },
+        data: { selectedListingId: listingId, edit: !editing },
       });
     }
 
@@ -77,20 +75,8 @@ export default ({ listingId = '', edit = false, createNewListing = false }) => {
   if (listingLoading || currentUserLoading) return 'Loading...';
   if (listingError || currentUserError) return `Error! ${listingError.message || currentUserError.message}`;
 
-  if (createNewListing) {
-    return (
-      <div className="card">
-        <EditListingForm handleToggleEditMode={toggleEditMode} editing={editing} id={listingData.listing.id} />
-      </div>
-    );
-  }
-
   if (editing) {
-    return (
-      <div className="card">
-        <EditListingForm handleToggleEditMode={toggleEditMode} editing={editing} {...listingData.listing} />
-      </div>
-    );
+    return <ProcessListingForm handleToggleEditMode={toggleEditMode} editing={editing} {...listingData.listing} />;
   }
 
   return (
